Added test_user
